dialogs moved, structurized abit
